agents/product-catalog-ad-generation/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "video-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.10.0",
    "google-cloud-aiplatform>=1.108.0",
    "google-cloud-bigquery>=3.35.1",
    "google-genai>=1.29.0",
    "langchain>=0.3.27",
    "pandas>=2.3.1",
    "toolbox-core>=0.4.0",
    "moviepy>=2.2.1",
    "google-cloud-texttospeech>=2.29.0",
    "soundfile>=0.13.1",
    "pytest>=8.4.2",
    "aiohttp>=3.12.15",
]

[tool.setuptools]
packages = ["content_gen_agent"]

[project.optional-dependencies]
dev = [
    "flake8>=7.3.0",
    "python-dotenv>=1.1.1",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.7",
]
